Organize code and add comments to new public properties
