+++ src/main/java/org/gradlex/javamodule/dependencies/initialization/Directory.java
@@ -54,8 +54,19 @@ public abstract class Directory {
     @Inject
     protected abstract ObjectFactory getObjects();
 
+    /**
+     * Configure which folders should be ignored when searching for Modules.
+     * This can be tweaked to optimize the configuration cache hit ratio.
+     * Defaults to: 'build', '.*'
+     */
     public abstract ListProperty<String> getExclusions();
 
+    /**
+     * Configure if only folders that contain a 'build.gradle' or 'build.gradle.kts'
+     * should be considered when searching for Modules.
+     * Setting this to true may improve configuration cache hit ratio if you know
+     * that all modules have build files in addition to the 'module-info.java' files.
+     */
     public abstract Property<Boolean> getRequiresBuildFile();
 
     /**
