+++ src/main/java/org/gradlex/javamodule/dependencies/internal/utils/ValueModuleDirectoryListing.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradlex.javamodule.dependencies.initialization;
+package org.gradlex.javamodule.dependencies.internal.utils;
 
 import org.gradle.api.provider.Property;
 import org.gradle.api.provider.SetProperty;
@@ -22,61 +22,43 @@ import org.gradle.api.provider.ValueSource;
 import org.gradle.api.provider.ValueSourceParameters;
 
 import java.io.File;
-import java.io.FileFilter;
 import java.io.IOException;
-import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.attribute.BasicFileAttributes;
-import java.util.Arrays;
 import java.util.List;
-import java.util.function.BiPredicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-public abstract class ValueSourceDirectoryListing implements ValueSource<List<String>, ValueSourceDirectoryListing.DirectoryListingParameter> {
+public abstract class ValueModuleDirectoryListing implements ValueSource<List<String>, ValueModuleDirectoryListing.Parameter> {
 
+    public interface Parameter extends ValueSourceParameters {
+        Property<File> getDir();
+        SetProperty<String> getExplicitlyConfiguredFolders();
+        SetProperty<String> getExclusions();
+        Property<Boolean> getRequiresBuildFile();
+    }
 
     @Override
     public List<String> obtain() {
         Path path = getParameters().getDir().get().toPath();
-        File file = getParameters().getDir().get();
-        try (Stream<Path> directoryStream = Files.find(path, 1, new BiPredicate<Path, BasicFileAttributes>() {
-            @Override
-            public boolean test(Path path, BasicFileAttributes basicFileAttributes) {
-                return basicFileAttributes.isDirectory();
-            }
-        })) {
-            return directoryStream.filter(x -> !getParameters().getExclusions().get().contains(x.getFileName().toString()))
-                    .filter(x -> getParameters().getRegexExclusions().get().stream().noneMatch(r -> x.getFileName().toString().matches(r)))
+        try (Stream<Path> directoryStream = Files.find(path, 1, (unused, basicFileAttributes) -> basicFileAttributes.isDirectory())) {
+            return directoryStream
+                    .filter(x -> !getParameters().getExplicitlyConfiguredFolders().get().contains(x.getFileName().toString()))
+                    .filter(x -> getParameters().getExclusions().get().stream().noneMatch(r -> x.getFileName().toString().matches(r)))
                     .filter(x -> checkBuildFile(x, getParameters()))
                     .map(x -> x.getFileName().toString())
                     .sorted()
                     .collect(Collectors.toList());
 
         } catch (IOException e) {
-            throw new RuntimeException("Failed on " + file, e);
+            throw new RuntimeException("Failed to inspect: " + path, e);
         }
-
-
     }
 
-    private boolean checkBuildFile(Path x, DirectoryListingParameter parameters) {
+    private boolean checkBuildFile(Path x, Parameter parameters) {
         if (!parameters.getRequiresBuildFile().get()) {
             return true;
         }
         return Files.isRegularFile(x.resolve("build.gradle.kts")) || Files.isRegularFile(x.resolve("build.gradle"));
-    }
-
-    interface DirectoryListingParameter extends ValueSourceParameters {
-
-        Property<Boolean> getRequiresBuildFile();
-
-        Property<File> getDir();
-
-        SetProperty<String> getExclusions();
-
-        SetProperty<String> getRegexExclusions();
-
     }
 }
